+++ src/main/java/org/f5games/aiden_game/repositories/CharacterRepository.java
@@ -1,7 +1,8 @@
 package org.f5games.aiden_game.repositories;
 
+import org.f5games.aiden_game.models.Character;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CharacterRepository extends JpaRepository<Character, Integer>{
+public interface CharacterRepository extends JpaRepository<Character, Long>{
     
 }
